Add gRPC JWKS authentication interceptor (#8)
This patch adds an authentication interceptor that expects a JWT token
in the authorization header and validates it against a configured list
of JSON web key set files or URLs.

This authentication mechanism is disabled by default, and to enable it
is necessary to add the following options to the command line of the
server:

    --grpc-authn-type=jwks
    --grpc-authn-jwks-url=...

There are additional command line options to configure the connection to
the URL that contains the JWKS. For example, a pod running inside
Kubernetes cluster can use the following options to use the JWKS that
contains the keys used to sign the service account tokens:

    --grpc-authn-type=jwks
    --grpc-authn-jwks-url=https://kubernetes.default.svc/openid/v1/jwks
    --grpc-authn-jwks-ca-file=/run/secrets/kubernetes.io/serviceaccount/ca.crt
    --grpc-authn-jwks-token-file=/run/secrets/kubernetes.io/serviceaccount/token

This patch adds those options to the manifests that deploy to a
Kubernetes cluster. It also adds a new `client` service account that can
be used to obtain the required token. For example, to obtain the list of
templats using `grpcurl`:

    $ token=$(kubectl create token -n innabox client)
    $ grpcurl -insecure \
    -H "Authorization: Bearer ${token}" \
    fulfillment-api-innabox.apps.mycluster.com:443 \
    fulfillment.v1.ClusterTemplates/List
    {...}

Note that this doesn't add any authorization, only authentication: as
long as you provide a valid service account token you will be able to
use the API. Aurhorization will be introduced later.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>