Implement field mask
Currently update requests need to pass the complete representation of
the object. For example, when updating a template it is necessary to
pass both the title and the description, even if they aren't going to be
modified. Those fields that aren't passed will be cleared. To avoid that
this patch implements the `update_mask` field, so that the client can
specify which fields should be updated.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>