Add `filter` to private `List` methods (#40)
All the `List` methods need to have a `filter` parameter in the request
type, otherwise any request will fail because the generic server needs
it. I forgot to add these parameters in the patch that added the support
for the `filter` parameter. This patch fixes that.

To reduce the chaces of this happening again in the future the patch
also introduces unit tests for the servers.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>