Don't clear other spec fields when changing node requests
Currently the code that we use to update the `nodeRequests` of the
`ClusterOrder` Kubernetes object doesn't preserve other fields, in
particular it doesn't preserve the `templateID` field. As a result the
update is rejected by the Kubernetes API server, because that field is
mandatory. This patch fixes that code so that it only changes the
`nodeRequests` field.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>