Convert objects to JSON wihthout removing id and metadata
Currently when an object is converted to JSON in order to save it to the
'data' column we first remove the 'id' and 'metadata' fields because
they are saved in separate columns. This complicates things when there
are multiple metadata implementantions. To avoid that this patch adds a
new JSON encoder that works like the default one, but that can also
ignore a set of fields.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>