Add finalizers
This patch adds the concept of _finalizers_, copied from the Kubernetes
API.

The private representation of objects will have a set of finalizers in
the `metadata` field. The finalizers will not be part of the public API.

Currently when an object is deleted it isn't completely removed, instead
of that the `deletion_timestamp` is set to the current date but the
database row is preserved. After this patch the row will be deleted if
the object has no finalizers. The row will be also deleted if when the
finalizers of an object that already has a value in the
`deletion_timestamp` are removed.

When an object is completely removed its data will be automatically
copied to the corresponding `archived_...` table. Note that this patch
doesn't add any mechanism to access those archives from the API, or to
clean them up.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>