Add integration tests configuration
This patch adds support for using configuration variables to change the
behaviour of the integration tests. The first such variable is
`IT_KEEP_KIND`. When set to `true` (the default is `false`) the kind
cluster used to run the tests will not be deleted. This allows the
developer to inspect the cluster, in particular the logs of the services.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>