Full sync of clusters on hub changes (#94)
Currently a cluster is successfully created even if there is no hub yet.
Later, when a hub is added nothing happens inmediatelly, instead the
cluster will be re-evaluated during the next full sync, which happens
every hour. In order to reduce that time this patch changes the cluster
reconciler so that it will also watch for changes in hubs. When a new
hub is added the reconciler will run the full sync again.

Related: https://github.com/innabox/issues/issues/181

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>